pyproject.toml
[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "missing-function-docstring",
    "missing-module-docstring",
    "missing-class-docstring",
    "too-few-public-methods",
    "too-many-instance-attributes",
    "too-many-arguments",
    "too-many-locals",
    "inconsistent-return-statements",
    "no-else-return",
    "broad-except",
    "logging-fstring-interpolation",
]

[tool.pylint.format]
# Maximum number of characters on a single line.
max-line-length = 100

[tool.pylint.similarities]
# Maximum number of characters on a single line.
ignore-imports = "yes"

[tool.pylint.basic]
# Maximum number of characters on a single line.
good-names = [
    "f", "s", "d", "i", "j", "k", "setUp", "tearDown"
]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
allow_redefinition = true
warn_return_any = false
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "common.*",
    "price_monitoring.async_runner",
]
ignore_errors = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
omit = [
    "*tests*",
    "*/abstract_*.py",
    "*/__init__.py",
    "price_monitoring/async_runner.py",
    "price_monitoring/common.py",
    "price_monitoring/constants.py",
    "price_monitoring/tracer.py",
]

[tool.poetry]
name = "steam_csmoney_price_monitoring"
version = "0.1.0"
description = "The system for monitoring item prices between CS.MONEY and Steam market"
authors = ["Viktor Kozlov <soniclev@mail.ru>"]

[tool.poetry.dependencies]
# For Scipy compatibility
python = ">=3.9,<3.11"
aio-pika = "^7.2.0"
dataclasses-json = "^0.5.7"
marshmallow-dataclass = "^8.5.8"
marshmallow = "^3.16.0"
marshmallow-enum = "^1.5.1"
aioredis = "^2.0.1"
aioredis-lock = "^0.1.0"
aiohttp = "^3.8.1"
aiohttp-socks = "^0.7.1"
PySocks = "^1.7.1"
random-user-agent = "^1.0.1"
json-logging = "^1.3.0"
PyYAML = "^6.0"
aiogram = "^2.20"
aiozipkin = "^1.1.1"
orjson = "^3.7.2"
uvloop = { version = "^0.16.0", optional = true }

[tool.poetry.extras]
linux = ["uvloop"]

[tool.poetry.dev-dependencies]
aioresponses = "^0.7.3"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
coverage = "^6.4.1"
pylint = "^2.14.3"
fakeredis = "^1.8.1"
freezegun = "^1.2.1"
memory-profiler = "^0.60.0"
line-profiler = "^3.5.1"
line-profiler-pycharm = "^1.1.0"
mypy = "^0.961"
types-PyYAML = "^6.0.7"
python-dotenv = "^0.20.0"
seaborn = "^0.11.2"
matplotlib = "^3.5.2"
notebook = "^6.4.12"
black = "^22.3.0"
pre-commit = "^2.19.0"
sortedcontainers = "^2.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
